moved checkifgameshouldfinish to phasegame helpers
